requirements/local.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.4
django-ckedit==1.0
django-environ==0.9.0
django-environ-2==2.3.0
django-model-utils==4.3.1
easy-thumbnails==2.8.4
Pillow==9.3.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
